+++ CHANGES.rst
@@ -15,6 +15,9 @@
 - Remove hard dependency on Astropy. It is still required for testing, and for
   processing ASDF-in-FITS files. [#476]
 
+- Add command for extracting ASDF extension from ASDF-in-FITS file and
+  converting it to a pure ASDF file. [#477]
+
 1.4.0 (unreleased)
 ------------------